refactored Result and Message
