+++ src/IO.Swagger.V1RC03/Controllers/AssetAdministrationShellAPI.cs
@@ -18,6 +18,7 @@ using IO.Swagger.V1RC03.Attributes;
 
 using Microsoft.AspNetCore.Authorization;
 using IO.Swagger.V1RC03.Models;
+using IO.Swagger.V1RC03.ApiModel;
 
 namespace IO.Swagger.V1RC03.Controllers
 {
