+++ src/IO.Swagger.V1RC03/Controllers/ISubmodelRepositoryAPI.cs
@@ -12,7 +12,6 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System.ComponentModel.DataAnnotations;
-using IO.Swagger.V1RC03.Models;//remove later
 using AasCore.Aas3_0_RC02;
 
 namespace IO.Swagger.V1RC03.Controllers
