+++ src/IO.Swagger.V1RC03/Controllers/SubmodelRepositoryAPI.cs
@@ -17,7 +17,7 @@ using System.ComponentModel.DataAnnotations;
 using IO.Swagger.V1RC03.Attributes;
 
 using Microsoft.AspNetCore.Authorization;
-using IO.Swagger.V1RC03.Models;//remove later
+using IO.Swagger.V1RC03.ApiModel;
 using AasCore.Aas3_0_RC02;
 
 namespace IO.Swagger.V1RC03.Controllers
