--- src/IO.Swagger.V1RC03/Models/Result.cs
@@ -1,135 +1,0 @@
-/*
- * DotAAS Part 2 | HTTP/REST | Entire API Collection
- *
- * The entire API collection as part of Details of the Asset Administration Shell Part 2
- *
- * OpenAPI spec version: V1.0RC03
- * 
- * Generated by: https://github.com/swagger-api/swagger-codegen.git
- */
-using System;
-using System.Linq;
-using System.IO;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
-using System.Runtime.Serialization;
-using Newtonsoft.Json;
-
-namespace IO.Swagger.V1RC03.Models
-{ 
-    /// <summary>
-    /// 
-    /// </summary>
-    [DataContract]
-    public partial class Result : IEquatable<Result>
-    { 
-        /// <summary>
-        /// Gets or Sets Messages
-        /// </summary>
-
-        [DataMember(Name="messages")]
-        public List<Message> Messages { get; set; }
-
-        /// <summary>
-        /// Gets or Sets Success
-        /// </summary>
-
-        [DataMember(Name="success")]
-        public bool? Success { get; set; }
-
-        /// <summary>
-        /// Returns the string presentation of the object
-        /// </summary>
-        /// <returns>String presentation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder();
-            sb.Append("class Result {\n");
-            sb.Append("  Messages: ").Append(Messages).Append("\n");
-            sb.Append("  Success: ").Append(Success).Append("\n");
-            sb.Append("}\n");
-            return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
-        }
-
-        /// <summary>
-        /// Returns true if objects are equal
-        /// </summary>
-        /// <param name="obj">Object to be compared</param>
-        /// <returns>Boolean</returns>
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((Result)obj);
-        }
-
-        /// <summary>
-        /// Returns true if Result instances are equal
-        /// </summary>
-        /// <param name="other">Instance of Result to be compared</param>
-        /// <returns>Boolean</returns>
-        public bool Equals(Result other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-
-            return 
-                (
-                    Messages == other.Messages ||
-                    Messages != null &&
-                    Messages.SequenceEqual(other.Messages)
-                ) && 
-                (
-                    Success == other.Success ||
-                    Success != null &&
-                    Success.Equals(other.Success)
-                );
-        }
-
-        /// <summary>
-        /// Gets the hash code
-        /// </summary>
-        /// <returns>Hash code</returns>
-        public override int GetHashCode()
-        {
-            unchecked // Overflow is fine, just wrap
-            {
-                var hashCode = 41;
-                // Suitable nullity checks etc, of course :)
-                    if (Messages != null)
-                    hashCode = hashCode * 59 + Messages.GetHashCode();
-                    if (Success != null)
-                    hashCode = hashCode * 59 + Success.GetHashCode();
-                return hashCode;
-            }
-        }
-
-        #region Operators
-        #pragma warning disable 1591
-
-        public static bool operator ==(Result left, Result right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(Result left, Result right)
-        {
-            return !Equals(left, right);
-        }
-
-        #pragma warning restore 1591
-        #endregion Operators
-    }
-}
